src/main/java/com/kkm/store/orders/controller/OrderController.java
package com.kkm.store.orders.controller;

import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

import com.kkm.store.orders.service.OrderService;

import java.util.Map;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;

@RestController
@RequestMapping("/api/orders")
public class OrderController {

    @Autowired
    private OrderService orderService;

    @GetMapping("/{orderId}")
    public Map<String, Object> getOrderById(@PathVariable String orderId) {
        return orderService.getOrderById(orderId);
    }
}
